# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "مسیر کانکور — API",
    "description": "API بک‌اند سایت آموزشگاه کانکور. اکثر مسیرها نیازمند توکن JWT هستند (هدر Authorization: Bearer <token>).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "نسبی — همان دامنه‌ی فرانت‌اند (پشت پروکسی/nginx)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
...
      "Video": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "duration_seconds": { "type": "integer" },
          "sort_order": { "type": "integer" }
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "description": "نقش کاربر اجازه‌ی این عملیات را نمی‌دهد.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "منبع مورد نظر یافت نشد.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/videos/{id}": {
      "get": {
        "tags": ["videos"],
        "summary": "جزئیات ویدیو + پیشرفت تماشا (نیازمند دسترسی به کورس)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Video" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/videos/upload/{courseId}": {
      "post": {
        "tags": ["videos"],
        "summary": "آپلود ویدیوی جدید برای یک کورس (استاد/ادمین)",
        "description": "اگر BUNNY_STREAM_LIBRARY_ID و BUNNY_STREAM_API_KEY تنظیم شده باشند، ویدیو مستقیم به Bunny Stream آپلود و url یک آدرس embed از iframe.mediadelivery.net می‌شود. در غیر این صورت روی دیسک محلی سرور (backend/uploads) ذخیره می‌شود.",
        "parameters": [{ "name": "courseId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "video": { "type": "string", "format": "binary" },
                  "title": { "type": "string" },
                  "sort_order": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "آپلود شد" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "502": { "description": "خطا در آپلود به Bunny Stream" }
        }
      }
    },
    "/videos/{id}/progress": {
      "post": {
        "tags": ["videos"],
        "summary": "ثبت پیشرفت تماشای ویدیو (فقط شاگرد)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "watched_seconds": { "type": "integer" }, "completed": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "ثبت شد" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
  }
}